New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s): sd            Description: added a line to check if temp_prefix exists.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@29501 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ gcwrap/python/scripts/task_sdsave.py
@@ -133,7 +133,8 @@ class sdsave_worker(sdutil.sdtask_template):
         if hasattr(self,'restore') and self.restore:
             casalog.post( "Restoring MOLECULE_ID column in %s "%self.infile )
             self.original_scan._setmolidcol_list(self.molids)
-        os.system('rm -rf %s*' % self.temp_prefix)
+        if hasattr(self,'temp_prefix') and self.splitant:
+            os.system('rm -rf %s*' % self.temp_prefix)
 
     def __dochannelrange(self, scantab):
         if len(self.spw) > 0:
